requirements.txt
# LandmarkFusion-ST Requirements
# Tested with Python 3.10+, CUDA 12.x, RTX 5060 8GB

# Core deep learning
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Landmark extraction
mediapipe>=0.10.9
opencv-python>=4.9.0

# Data / ML utilities
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
tqdm>=4.66.0
Pillow>=10.2.0

# Semantic refinement (BERT)
transformers>=4.38.0
sentencepiece>=0.1.99

# Logging / visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0
tensorboard>=2.16.0
